creating an item delay

+++ Assets/Source/Inventory/Scripts/Item/ItemData/FoodItemData.cs
@@ -3,13 +3,13 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Inventory System/Item Data/Food", order = 51)]
 public class FoodItemData : InventoryItemData
 {
-    [SerializeField] private int _amountWater;
-    [SerializeField] private int _amountSatiety;
-    [SerializeField] private int _amountHealth;
+    [SerializeField] private float _amountWater;
+    [SerializeField] private float _amountSatiety;
+    [SerializeField] private float _amountHealth;
 
-    public int AmountWater => _amountWater;
-    public int AmountSatiety => _amountSatiety;
-    public int AmountHealth => _amountHealth;
+    public float AmountWater => _amountWater;
+    public float AmountSatiety => _amountSatiety;
+    public float AmountHealth => _amountHealth;
 
     public void Eat()
     {
